pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "selftest"
version = "0.1.0"
description = "Auto-generate and run self-tests for Python scripts"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "rich>=13.0",
    "jinja2>=3.0",
    "pyyaml>=6.0",
    "openai>=1.0",
    "requests>=2.28",
]

[project.scripts]
selftest = "selftest.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
